POS-Texano/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy>=2.0.36
pymysql>=1.1.1
cryptography>=44.0.0
alembic>=1.14.1
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.2.0
python-multipart>=0.0.20
aiofiles>=24.1.0
Pillow>=11.1.0
email-validator>=2.2.0



